requirements.txt
line-bot-sdk==1.7.2
flask==1.0.2
requests==2.19.1
gunicorn==19.8.1
pylint==1.9.2
codecov==2.0.15
python-dateutil==2.7.3